Abstractのstyleを統合
